refactor: Move getTransferBuffers into the uhk-usb package
